move emergency access out of core folder - add more organization to components under settings
