+++ apps/web/src/app/auth/core/services/index.ts
@@ -1,2 +1,1 @@
-export * from "./emergency-access";
 export * from "./webauthn-login";
