+++ apps/web/src/app/auth/emergency-access/emergency-access.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from "@angular/core";
 
-import { EmergencyAccessApiService } from "./emergency-access-api.service";
-import { EmergencyAccessService } from "./emergency-access.service";
+import { EmergencyAccessApiService } from "./services/emergency-access-api.service";
+import { EmergencyAccessService } from "./services/emergency-access.service";
 
 @NgModule({
   declarations: [],
