+++ apps/web/src/app/auth/emergency-access/index.ts
@@ -1,0 +1,2 @@
+export * from "./emergency-access.module";
+export * from "./services";
