+++ apps/web/src/app/auth/emergency-access/services/index.ts
@@ -1,2 +1,1 @@
-export * from "./emergency-access.module";
 export * from "./emergency-access.service";
