+++ apps/web/src/app/auth/settings/emergency-access/confirm/emergency-access-confirm.component.ts
