+++ apps/web/src/app/auth/settings/settings.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from "@angular/core";
 import { PasswordCalloutComponent } from "@bitwarden/auth";
 
 import { SharedModule } from "../../shared";
-import { EmergencyAccessModule } from "../core";
+import { EmergencyAccessModule } from "../emergency-access";
 
 import { ChangePasswordComponent } from "./change-password.component";
 import { WebauthnLoginSettingsModule } from "./webauthn-login-settings";
